Fix for IndexOutOfRange on Unmatched loop tokens
Some invalid Brainfuck code with an unequal amount loop start/end tokens will lead to an IndexOutOfRange exception. This fix will ensure that the program does not attempt to traverse beyond the code range and will stop execution silently instead.

Some implementations of BF refuse to run the code immediately upon detecting it as invalid. This currently does not perform any pre-run checks, and as such some portions of the code may finish executing by the time the loop is encountered.